feat: _internal and lazyloading ml frameworks
