+++ tests/integration/frameworks/test_keras_impl.py
@@ -1,10 +1,12 @@
+import typing as t
+from typing import TYPE_CHECKING
+
 import numpy as np
 import psutil
 import pytest
 import tensorflow as tf
 import tensorflow.keras as keras
 
-import bentoml.keras
 from tests.utils.frameworks.tensorflow_utils import (
     CustomLayer,
     KerasSequentialModel,
@@ -12,14 +14,22 @@ from tests.utils.frameworks.tensorflow_utils import (
 )
 from tests.utils.helpers import assert_have_file_extension
 
-TF2 = tf.__version__.startswith("2")
+if TYPE_CHECKING:
+    from bentoml._internal.models import ModelStore
+
+try:
+    import importlib.metadata as importlib_metadata
+except ImportError:
+    import importlib_metadata
+
+TF2 = importlib_metadata.version("tensorflow").startswith("2")
 MODEL_NAME = __name__.split(".")[-1]
 
 test_data = [1, 2, 3, 4, 5]
 res = KerasSequentialModel().predict(np.array([test_data]))
 
 
-def predict_assert_equal(model: keras.Model):
+def predict_assert_equal(model: "keras.Model") -> None:
     t_data = np.array([test_data])
     assert model.predict(t_data) == res
 
@@ -41,7 +51,9 @@ def predict_assert_equal(model: keras.Model):
         ),
     ],
 )
-def test_keras_save_load(model, kwargs, modelstore):
+def test_keras_save_load(
+    model: "keras.Model", kwargs: t.Dict[str, t.Any], modelstore: "ModelStore"
+) -> None:
     tag = bentoml.keras.save(MODEL_NAME, model, **kwargs, model_store=modelstore)
     model_info = modelstore.get(tag)
     if kwargs["custom_objects"] is not None:
@@ -63,7 +75,7 @@ def test_keras_save_load(model, kwargs, modelstore):
 
 
 @pytest.mark.skipif(not TF2, reason="Tests for Tensorflow 2.x")
-def test_keras_v2_setup_run_batch(modelstore):
+def test_keras_v2_setup_run_batch(modelstore: "ModelStore") -> None:
     model_class = KerasSequentialModel()
     tag = bentoml.keras.save(MODEL_NAME, model_class, model_store=modelstore)
     runner = bentoml.keras.load_runner(tag, model_store=modelstore)
@@ -75,17 +87,16 @@ def test_keras_v2_setup_run_batch(modelstore):
 
 
 @pytest.mark.skipif(TF2, reason="Tests for Tensorflow 1.x")
-def test_keras_v1_setup_run_batch(modelstore):
+def test_keras_v1_setup_run_batch(modelstore: "ModelStore") -> None:
     model_class = KerasSequentialModel()
     tag = bentoml.keras.save(MODEL_NAME, model_class, model_store=modelstore)
     runner = bentoml.keras.load_runner(tag, model_store=modelstore)
     with runner._session.as_default():
-
         assert runner.run_batch([test_data]) == res
 
 
 @pytest.mark.gpus
-def test_tensorflow_v2_setup_on_gpu(modelstore):
+def test_tensorflow_v2_setup_on_gpu(modelstore: "ModelStore") -> None:
     model_class = KerasSequentialModel()
     tag = bentoml.keras.save(MODEL_NAME, model_class, model_store=modelstore)
     runner = bentoml.keras.load_runner(tag, model_store=modelstore)
